v1.24 SD-98,SD-97,SD-99,SD-95,SD-96

+++ console/controllers/CrudController.php
@@ -31,49 +31,62 @@ class CrudController extends Controller
 
     public function actionProcessOrderQuery()
     {
-        for ($i = 0; $i < 5; $i++) {
+        for ($i = 0; $i < 6; $i++) {
             $data = OrderQuery::find()
                 ->where(['order' => 0])
                 ->all();
 
-            print("START PROCESS\n");
-            print("QUERY ITEMS: " . count($data) ."\n");
-            foreach ($data as $item) {
-                $supplierId = $item->supplier_id;
-                $orderId = $item->order_id;
+            print("START PROCESS\n " . date("H:i:s d-m-Y") . "\n");
 
-                if ($item->time_start_query == 0) {
-                    print("NEXT SUPPLIER\n");
-                    $item->time_start_query = time();
-                    $item->save();
-
-                    $order = Order::findOne($orderId);
-
-                    $messageSupplier = "You have a new order available for the next 2 minutes until " . date ('g:i:s A', time() + 120) ." Click to View/Accept/Decline this order " . \Yii::$app->params['webProjectUrl'] . '/supplier/show-order?l=' . $order->weblink;
-
-                    $supplier = Supplier::find()->where(['id' => $supplierId])->one();
-                    $supplierUser = User::find()->where(['id' => $supplier->supplier_id])->one();
-
-                    Twilio::sendSms($supplierUser->phone_number, $messageSupplier);
+            if (!count($data)) {
+                print("NO ORDERS IN PROGRESS\n");
+                foreach(Order::findAll(['status' => Order::ORDER_STATUS_NEW]) as $order) {
+                    $user = User::findOne($order->customer_id);
+                    $messageCustomer = "Unfortunatly, nobody takes your order. Please, try again later.";
+                    Twilio::sendSms($user->phone_number, $messageCustomer);
+                    $order->status = Order::ORDER_STATUS_CANCELLED;
+                    $order->save();
                 }
+            } else {
+                print("QUERY ITEMS: " . count($data) ."\n");
+                foreach ($data as $item) {
+                    $supplierId = $item->supplier_id;
+                    $orderId = $item->order_id;
 
-                if (($item->time_start_query + \Yii::$app->params['timeToTakeOrder']) <= time()) {
-                    $item->delete();
-                    print("EXPIRED\n");
+                    if ($item->time_start_query == 0) {
+                        print("NEXT SUPPLIER\n");
+                        $item->time_start_query = time();
+                        $item->save();
 
-                    if (!OrderQuery::find()->where(['order_id' => $orderId])->count()) {
                         $order = Order::findOne($orderId);
-                        $user = User::findOne($order->customer_id);
-                        $messageCustomer = "Unfortunatly, nobody takes your order. Please, try again later.";
-                        Twilio::sendSms($user->phone_number, $messageCustomer);
-                        Order::updateAll(['status' => Order::ORDER_STATUS_CANCELLED], ['id' => $orderId]);
-                    } else {
-                        OrderQuery::updateAllCounters(['order' => -1],['order_id' => $orderId]);
+
+                        $messageSupplier = "You have a new order available for the next 2 minutes until " . date ('g:i:s A', time() + 120) ." Click to View/Accept/Decline this order " . \Yii::$app->params['webProjectUrl'] . '/supplier/show-order?l=' . $order->weblink;
+
+                        $supplier = Supplier::find()->where(['id' => $supplierId])->one();
+                        $supplierUser = User::find()->where(['id' => $supplier->supplier_id])->one();
+
+                        Twilio::sendSms($supplierUser->phone_number, $messageSupplier);
+                    }
+
+                    if (($item->time_start_query + \Yii::$app->params['timeToTakeOrder']) <= time()) {
+                        $item->delete();
+                        print("EXPIRED\n");
+
+                        if (!OrderQuery::find()->where(['order_id' => $orderId])->count()) {
+                            $order = Order::findOne($orderId);
+                            $user = User::findOne($order->customer_id);
+                            $messageCustomer = "Unfortunatly, nobody takes your order. Please, try again later.";
+                            Twilio::sendSms($user->phone_number, $messageCustomer);
+                            Order::updateAll(['status' => Order::ORDER_STATUS_CANCELLED], ['id' => $orderId]);
+                        } else {
+                            OrderQuery::updateAllCounters(['order' => -1],['order_id' => $orderId]);
+                        }
                     }
                 }
             }
-
-            sleep(10);
+            if ($i < 6) {
+                sleep(10);
+            }
         }
     }
 }
